Integer ==> xsd:long
git-svn-id: https://src.springsource.com/svn/wm_svn/eng3/trunk@32737 bab4cc61-abf4-0310-9601-8b162d8b0fb1

+++ dev/core/tools/src/com/wavemaker/tools/ws/infoteria/FlowSupport.java
@@ -389,7 +389,7 @@ public class FlowSupport {
                 rtn = "string";
                 break;
             case Integer:
-                rtn = "int";
+                rtn = "long";
                 break;
             case Double:
                 rtn = "double";
